# backend-client/.openapi-source/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "shepard backend",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/shepard/api",
    "description" : "shepard backend"
  } ],
  "security" : [ {
    "bearer" : [ ]
  }, {
    "apikey" : [ ]
  } ],
...
  "paths" : {
...
    "/spatialDataContainers/{spatialDataContainerId}/payload" : {
      "get" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Get spatial data by container id",
        "operationId" : "getSpatialDataPoints",
        "parameters" : [ {
          "name" : "metadataFilter",
          "in" : "query",
          "description" : "This filter should be a stringified list of JSON object for exact match in metadata. \n\nExample : `{\"track\": 1, \"layer\": 4, \"key\": {\"subKey\": \"some data\"}}`\n",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementsFilter",
          "in" : "query",
          "description" : "This filter should be a stringified list of JSON FilterConditions. \n\nFilterCondition has this structure: `{\"key\": <KEY>, \"operator\": <OPERATOR>, \"value\": <VALUE>}`. \n\nThe key is a comma separated list of keys representing the path to the value. \n\nThe operator is one of `EQUALS`, `GREATER_THAN` or `LESS_THAN`. \n\nThe value needs to be a number. \n\nExample: `[{\"key\":\"temperature,val\",\"operator\":\"EQUALS\",\"value\":20},{\"key\":\"temperature,val\",\"operator\":\"LESS_THAN\",\"value\":10}]`\n",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "geometryFilter",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "examples" : {
            "K Nearest Neighbor" : {
              "value" : {
                "type" : "K_NEAREST_NEIGHBOR",
                "k" : 5,
                "x" : 10,
                "y" : 20,
                "z" : 30
              }
            },
            "Bounding Box" : {
              "value" : {
                "type" : "AXIS_ALIGNED_BOUNDING_BOX",
                "minX" : 0,
                "minY" : 0,
                "minZ" : 0,
                "maxX" : 100,
                "maxY" : 100,
                "maxZ" : 100
              }
            },
            "Bounding Sphere" : {
              "value" : {
                "type" : "BOUNDING_SPHERE",
                "radius" : 50,
                "centerX" : 15,
                "centerY" : 25,
                "centerZ" : 20
              }
            }
          }
        }, {
          "name" : "startTime",
          "in" : "query",
          "description" : "Start timestamp in nanoseconds, exclusive",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "endTime",
          "in" : "query",
          "description" : "End timestamp in nanoseconds, exclusive",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "Returns every nth data point from the container. We use the modulo operator on the point ids, therefore an even distribution cannot be guaranteed.",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SpatialDataPoint"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      },
      "post" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Adds spatial data points to a spatial data container.",
        "operationId" : "createSpatialDataPoints",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/SpatialDataPoint"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      }
    },
...
    "/spatialDataContainers/{spatialDataContainerId}" : {
      "get" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Get spatial data container.",
        "operationId" : "getSpatialDataContainer",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SpatialDataContainer"
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Deletes spatial data container and related spatial data.",
        "operationId" : "deleteSpatialDataContainer",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok"
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      }
    },
...
    "/spatialDataContainers" : {
      "get" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Get spatial data containers.",
        "operationId" : "getSpatialDataContainers",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "orderBy",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/ContainerAttributes"
          }
        }, {
          "name" : "orderDesc",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SpatialDataContainer"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "404" : {
            "description" : "not found"
          }
        }
      },
      "post" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Create a new spatial data container.",
        "operationId" : "createSpatialDataContainer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SpatialDataContainer"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SpatialDataContainer"
                }
              }
            }
          }
        }
      }
    },
...
    "/spatialDataContainers/{spatialDataContainerId}/permissions" : {
      "get" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Get permissions",
        "operationId" : "getSpatialDataPermissions",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Permissions"
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      },
      "put" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Edit permissions",
        "operationId" : "editSpatialDataPermissions",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Permissions"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Permissions"
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      }
    },
...
    "/spatialDataContainers/{spatialDataContainerId}/roles" : {
      "get" : {
        "tags" : [ "spatialDataContainer" ],
        "description" : "Get roles",
        "operationId" : "getSpatialDataRoles",
        "parameters" : [ {
          "name" : "spatialDataContainerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Roles"
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "401" : {
            "description" : "not authorized"
          },
          "403" : {
            "description" : "forbidden"
          },
          "404" : {
            "description" : "not found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Permissions" : {
        "required" : [ "entityId", "permissionType", "reader", "writer", "manager" ],
        "type" : "object",
        "properties" : {
          "entityId" : {
            "format" : "int64",
            "type" : "integer",
            "readOnly" : true
          },
          "owner" : {
            "type" : "string"
          },
          "permissionType" : {
            "$ref" : "#/components/schemas/PermissionType"
          },
          "reader" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "writer" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "readerGroupIds" : {
            "type" : "array",
            "items" : {
              "format" : "int64",
              "type" : "integer"
            }
          },
          "writerGroupIds" : {
            "type" : "array",
            "items" : {
              "format" : "int64",
              "type" : "integer"
            }
          },
          "manager" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "SpatialDataContainer" : {
        "required" : [ "id", "createdAt", "createdBy", "updatedAt", "updatedBy", "name", "type" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer",
            "readOnly" : true
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string",
            "readOnly" : true,
            "example" : "2024-08-15T11:18:44.632+00:00"
          },
          "createdBy" : {
            "type" : "string",
            "readOnly" : true
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string",
            "readOnly" : true,
            "example" : "2024-08-15T11:18:44.632+00:00",
            "nullable" : true
          },
          "updatedBy" : {
            "type" : "string",
            "readOnly" : true,
            "nullable" : true
          },
          "name" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "allOf" : [ {
              "$ref" : "#/components/schemas/ContainerType"
            } ],
            "readOnly" : true
          }
        }
      },
...
      "SpatialDataPoint" : {
        "required" : [ "x", "y", "z", "measurements", "metadata" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "format" : "int64",
            "description" : "Time in nanoseconds since unix epoch. If no value is provided, the current server time is used.",
            "type" : "integer"
          },
          "x" : {
            "format" : "double",
            "description" : "X Coordinate",
            "type" : "number"
          },
          "y" : {
            "format" : "double",
            "description" : "Y Coordinate",
            "type" : "number"
          },
          "z" : {
            "format" : "double",
            "description" : "Z Coordinate",
            "type" : "number"
          },
          "measurements" : {
            "description" : "measurement",
            "type" : "object",
            "additionalProperties" : { }
          },
          "metadata" : {
            "description" : "Data point meta data",
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
      "Roles" : {
        "required" : [ "owner", "manager", "writer", "reader" ],
        "type" : "object",
        "properties" : {
          "owner" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "manager" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "writer" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "reader" : {
            "type" : "boolean",
            "readOnly" : true
          }
        }
      },
...
      "PermissionType" : {
        "enum" : [ "Public", "PublicReadable", "Private" ],
        "type" : "string"
      },
...
      "ContainerAttributes" : {
        "enum" : [ "createdAt", "updatedAt", "name" ],
        "type" : "string"
      },
...
      "ContainerType" : {
        "enum" : [ "FILE", "TIMESERIES", "STRUCTUREDDATA", "SPATIALDATA", "BASIC" ],
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apikey" : {
        "type" : "apiKey",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    }
  }
}